fix: use main.ts and change webpack.config.js: - finally set up webpack and lambda configs 667!
