+++ nest-cart-service/tsconfig.json
@@ -6,6 +6,7 @@
       "es2020",
       "dom"
     ],
+    "esModuleInterop": true,
     "declaration": true,
     "strict": true,
     "noImplicitAny": true,
