+++ webpack.config.js
@@ -7,6 +7,10 @@ module.exports = (options, webpack) => {
   return {
     ...options,
     externals: [],
+    output: {
+      ...options.output,
+      libraryTarget: 'commonjs2',
+    },
     plugins: [
       ...options.plugins,
       new webpack.IgnorePlugin({
